requirements.txt
streamlit
torch
transformers
Pillow
opencv-python
pyttsx3
